Full UI version 3.03

+++ app/Http/Controllers/CategoryControllerWithRepos.php
@@ -89,6 +89,18 @@ class CategoryControllerWithRepos extends Controller
         return redirect()->action('CategoryControllerWithRepos@index')
             ->with('msg', 'Update Successfully');;
     }
+////    //        CategoryRepos::delete($id);
+//        $product = CategoryRepos::getCategoriesById($id);
+//
+//        if (count($product) > 0) {
+//            return redirect()->action('CategoryControllerWithRepos@index')
+//                ->with('msg', 'Can not delete!!! Please check animals before delete category!!!');
+//        } else {
+//            CategoryRepos::delete($id);
+//            return redirect()->action('CategoryControllerWithRepos@index')
+//                ->with('msg', 'Delete Successfully');
+//        }
+//
 
     public function confirm($id_cate){
         $category = CategoryShowRepos::getCategoryById($id_cate); //this is always an array
@@ -106,12 +118,25 @@ class CategoryControllerWithRepos extends Controller
 
             return redirect()->action('CategoryControllerWithRepos@index');
         }
+        $product = CategoryShowRepos::getCategoryById($id_cate);
 
-        CategoryShowRepos::delete($id_cate);
+//        CategoryShowRepos::delete($id_cate);
 
+        if (count($product) > 0) {
+            return redirect()->action('CategoryControllerWithRepos@index')
+                ->with('msg', 'Can not delete!!! Please check animals before delete category!!!');
+        } else {
+            CategoryShowRepos::delete($id_cate);
+            return redirect()->action('CategoryControllerWithRepos@index')
+                ->with('msg', 'Delete Successfully');
+        }
 
-        return redirect()->action('CategoryControllerWithRepos@index')
-            ->with('msg', 'Delete Successfully');
+//        CategoryShowRepos::delete($id_cate);
+//
+//
+//        return redirect()->action('CategoryControllerWithRepos@index')
+//            ->with('msg', 'Delete Successfully');
+
     }
 
     private function formValidate($request)
